refactor + add isBeta prop to main layout
