+++ src/components/Navbar/index.tsx
@@ -1,44 +1,23 @@
-"use client";
 import { Branding } from "./Branding";
-import { ConnectButton } from "./ConnectButton";
-import { Items } from "./Items";
-import ThemeSwitcher from "./ThemeSwitcher";
-import { useMediaQuery } from "src/hooks/useMediaQuery";
-const desktop = "(min-width: 768px)";
+import { ClientNavbar } from "./ClientNavbar";
 
-export const Navbar = () => {
-  const isDesktop = useMediaQuery(desktop);
-
+// Server component that evaluates the beta flag
+export async function Navbar({ isBeta }: { isBeta: boolean }) {
   return (
-    <>
-      <div
-        className={`
-          z-10
-          flex w-full flex-row items-center justify-between gap-8
-          p-3 font-secondary
-        `}
-      >
-        <div className="inline-flex flex-row items-center">
-          <Branding />
-          {/* <ThemeSwitcher /> */}
-
-          {/* <span className="ml-2 hidden sm:block">
-            <div className="font-base inline-flex cursor-pointer flex-row items-center font-bitter italic text-slate-700">
-              <span className="hover:text-slate-500">share stories</span>
-              <Dot />
-              <span className="hover:text-slate-500">remember memories</span>
-              <Dot />
-              <span className="hover:text-slate-500">explore ideas</span>{" "}
-              <Dot />
-              <span className="hover:text-slate-500">learn together</span>
-            </div>
-          </span> */}
-        </div>
-        <div className="inline-flex flex-row items-center gap-4">
-          {/* <Items /> */}
-          <ConnectButton />
-        </div>
+    <div
+      className={`
+        z-10
+        flex w-full flex-row items-center justify-between gap-8
+        p-3 font-secondary
+      `}
+    >
+      <div className="inline-flex flex-row items-center">
+        {/* Pass the evaluated flag to the client component */}
+        <Branding isBeta={isBeta} />
       </div>
-    </>
+
+      {/* Client component for interactive elements */}
+      <ClientNavbar />
+    </div>
   );
-};
+}
